requirements.txt
python-telegram-bot==13.15
python-dotenv~=1.0.0
pillow>=10.0.0
urllib3<2.0

